ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
langchain-core==0.3.10
langchain-ollama==0.2.3
langgraph==0.2.28
faiss-cpu==1.8.0
tavily-python==0.5.0
beautifulsoup4==4.12.3
httpx==0.27.2
numpy<2
